Add MTA-STS policy file for configuring TLS-RPT in DNS
Setup a MTA-STS policy to ensure that TLS is always used, and to provide
a way to for sending servers to refuse to deliver messages to servers
that don’t support TLS and have a trusted certificate.

Jekyll does not serve directories that start with a dot unless
explicitly told, which is necessary for MTA-STS, thus add '.nojekyll' to
the repo.

https://emailsecurity.blog/hosting-your-mta-sts-policy-using-github-pages